Update API tests
- Add test to KeywordController.

- Update open api v3.

- Add test for open api v3.

+++ api/src/main/kotlin/edu/wgu/osmt/keyword/KeywordTaskProcessor.kt
@@ -33,7 +33,7 @@ class KeywordTaskProcessor {
     fun removeKeyword(task: RemoveKeywordTask) {
         logger.info("Started processing to remove keyword task id: ${task.keywordId}")
 
-        val batchResult = keywordRepository.remove(task.keywordId.toLong())
+        val batchResult = keywordRepository.remove(task.keywordId)
 
         taskMessageService.publishResult(
             task.copy(result=batchResult, status= TaskStatus.Ready)
